Support varargs
From now on, all JS functions have the arity 1 and take a list of
arguments of arbitrary length, which is available within JS functions
as the `arguments` variable.

Arguments that are specified in a JS function's signature but are
omitted when calling it are initialized with the value `undefined`.

At present, the `arguments` variable is not really useful, as there's
currently no way to retrieve the arguments from the `arguments`
variable yet. This has to wait until we support proper objects.